pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "miachat"
version = "0.1.0"
description = "An AI personality system for creating and managing chat personalities"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    # Core dependencies
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    
    # LLM providers
    "openai>=1.3.0",
    "anthropic>=0.5.0",
    "pyngrok>=6.0.0",
    
    # Web server
    "gunicorn>=21.2.0",
    
    # Memory and search
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.3",
    "scikit-learn>=1.3.0",
    "faiss-cpu>=1.7.4",
    
    # Document processing
    "PyMuPDF>=1.23.8",
    "markdown>=3.5.2",
    "python-dateutil>=2.8.2",
    "tiktoken>=0.5.2",
    
    # Serialization and validation
    "pyyaml>=6.0.1",
    "jsonschema>=4.17.3",
    "xmlschema>=2.3.1",
    
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.9",
    
    # Web interface
    "flask-socketio>=5.3.6",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.1",
    
    # Utilities
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.1",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.1",
    
    # Code Quality
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "flake8>=4.0.0",
    "flake8-docstrings>=1.6.0",
    "flake8-bugbear>=22.0.0",
    
    # Development Tools
    "ipython>=8.0.0",
    "ipdb>=0.13.0",
    "watchdog>=2.1.0",
    
    # Documentation
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0"
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src/miachat"

[tool.coverage.run]
source = ["src/miachat"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
    "@abstractmethod",
]

[project.scripts]
miachat = "miachat.cli:cli" 